requirements.txt
discord.py
yfinance
SQLAlchemy
